migrations/models/2_20250307073353_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "flower" ADD "img_link" VARCHAR(500) NOT NULL DEFAULT '';"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "flower" DROP COLUMN "img_link";"""
